refactor(DoPayRequestDto)
- 가격의 정확도를 높이기 위해 리팩토링 진행
- 결제금액, 부가가치세 금액을 BigDecimal 타입으로 변경

+++ src/main/java/com/kakao/preinterview/payment/domain/payment/PaymentFactory.java
@@ -34,7 +34,7 @@ public class PaymentFactory {
             Long cardNumber,
             String duration,
             Integer cvc,
-            Long taxAmount
+            BigDecimal taxAmount
     ) {
         CardInfo cardInfo = CardInfo.create(cardNumber, duration, cvc);
 
@@ -43,7 +43,7 @@ public class PaymentFactory {
                 null,
                 PayInfo.create(installmentMonths, payAmount, payType),
                 cardInfo,
-                Tax.manualCreate(BigDecimal.valueOf(taxAmount), payAmount)
+                Tax.manualCreate(taxAmount, payAmount)
         );
     }
